make top corner the profile component dynamic.

+++ frontend/Utils/APIManager.js
@@ -3,8 +3,10 @@ import { fetchWithToken } from "/root/fetchWithToken.js"
 
 export async function getApiData(APIUrl) {
     const accessToken = localStorage.getItem('accessToken');
-    if (!accessToken)
+    if (!accessToken) {
+        console.log("Access token is missing.");
         return null;
+    }
     const response = await fetchWithToken(APIUrl,
     {
         method: "GET",
@@ -25,6 +27,10 @@ export async function getApiData(APIUrl) {
 export async function createApiData(APIUrl, body) {
     
     const accessToken = localStorage.getItem('accessToken');
+    if (!accessToken) {
+        console.log("Access token is missing.");
+        return null;
+    }
     const response = await fetchWithToken(APIUrl,
     {
         method: "POST",
@@ -34,6 +40,8 @@ export async function createApiData(APIUrl, body) {
         },
         body: body,
     });
+    if (!response.ok)
+        return null;
     return await response.json();
 
 }
@@ -41,7 +49,7 @@ export async function createApiData(APIUrl, body) {
 export async function updateApiData(APIUrl, body) {
     const accessToken = localStorage.getItem('accessToken');
     if (!accessToken) {
-        console.error("Access token is missing.");
+        console.log("Access token is missing.");
         return null;
     }
     const response = await fetchWithToken(APIUrl,
@@ -52,12 +60,18 @@ export async function updateApiData(APIUrl, body) {
         },
         body: body,
     });
+    if (!response.ok)
+        return null;
     return await response.json();
 
 }
 
 export async function deleteApiData(APIUrl) {
     const accessToken = localStorage.getItem('accessToken');
+    if (!accessToken) {
+        console.log("Access token is missing.");
+        return null;
+    }
     const response = await fetchWithToken(APIUrl,
     {
         method: "DELETE",
@@ -66,5 +80,7 @@ export async function deleteApiData(APIUrl) {
             "Content-Type": "application/json"
         }
     });
+    if (!response.ok)
+        return null;
     return await response.json();
 }
